✨ Finish invitation feature
