+++ events/core/models/__init__.py
@@ -1,2 +1,3 @@
 from events.core.models.event import Event
-from events.core.models.user import CustomUser, Invitation
+from events.core.models.invitation import Invitation
+from events.core.models.user import CustomUser
