+++ events/core/serializers/__init__.py
@@ -1,1 +1,2 @@
 from events.core.serializers.event import DetailEventSerializer, ListEventSerializer
+from events.core.serializers.invitation import UpdateInvitationSerializer
