+++ events/core/serializers/invitation.py
@@ -1,0 +1,9 @@
+from rest_framework.serializers import ModelSerializer
+
+from events.core.models import Invitation
+
+
+class UpdateInvitationSerializer(ModelSerializer):
+    class Meta:
+        model = Invitation
+        fields = ["status"]
