+++ events/core/views/__init__.py
@@ -1,2 +1,3 @@
 from events.core.views.event import EventViewSet
+from events.core.views.invitation import InvitationViewSet
 from events.core.views.user import UserViewSet
