+++ tests/unit/test_user.py
@@ -1,7 +1,6 @@
 from django.test import TestCase
 
 from events.core.models import CustomUser
-from events.core.models.user import Invitation
 
 
 class UserTestCase(TestCase):
@@ -30,23 +29,3 @@ class UserTestCase(TestCase):
         self.assertEqual(hasattr(CustomUser, "bio"), True)
         self.assertEqual(hasattr(CustomUser, "profile_picture"), True)
         self.assertEqual(hasattr(CustomUser, "friends"), True)
-
-
-class InvitationTestCase(TestCase):
-    """
-    Test Invitation model that should has following attributes:
-        - type
-        - status
-        - invitation_from
-        - invitation_to
-        - created_at
-        - updated_at
-    """
-
-    def test_model_has_attributes(self):
-        self.assertEqual(hasattr(Invitation, "type"), True)
-        self.assertEqual(hasattr(Invitation, "status"), True)
-        self.assertEqual(hasattr(Invitation, "invitation_from"), True)
-        self.assertEqual(hasattr(Invitation, "invitation_to"), True)
-        self.assertEqual(hasattr(Invitation, "created_at"), True)
-        self.assertEqual(hasattr(Invitation, "updated_at"), True)
